ispit kontroler i ispit period kontroler

+++ projekatTSEO/src/main/java/projekatTSEO/repository/IspitRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekatTSEO.model.Ispit;
 
-public interface IspitRepository extends JpaRepository<Ispit, Integer> {
+public interface IspitRepository extends JpaRepository<Ispit, Long> {
 	
 
 }
